remove SubscriptionManager and move to FilterSDK
